feat: adjusting new sections after merging to main
