+++ frontend/src/components/SeedlotRegistrationSteps/ParentTreeStep/constants.tsx
@@ -837,3 +837,13 @@ export const defaultMeanGeomConfig: MeanGeomInfoSectionConfigType = {
     }
   }
 };
+
+export const MIN_NE = '0.1';
+
+export const MAX_NE = '999.9';
+
+export const MAX_NE_DECIMAL = 1;
+
+export const INVALID_NE_RANGE_MSG = `Must be between ${MIN_NE} and ${MAX_NE}`;
+
+export const INVALID_NE_DECIMAL_MSG = `Cannot have more than ${MAX_NE_DECIMAL} decimal place`;
