+++ frontend/src/components/SeedlotReviewSteps/Collection/Read/index.tsx
@@ -12,6 +12,7 @@ import { getForestClientLabel } from '../../../../utils/ForestClientUtils';
 import getConeCollectionMethod from '../../../../api-service/coneCollectionMethodAPI';
 import { THREE_HALF_HOURS, THREE_HOURS } from '../../../../config/TimeUnits';
 import { formatCollectionMethods } from '../utils';
+import GeoInfo from '../GeoInfo';
 
 const CollectionReviewRead = () => {
   const {
@@ -141,6 +142,10 @@ const CollectionReviewRead = () => {
           />
         </Column>
       </Row>
+
+      <Divider />
+
+      <GeoInfo />
     </FlexGrid>
   );
 };
