--- frontend/src/components/SeedlotReviewSteps/Collection/constants.ts
@@ -1,9 +1,0 @@
-export const MIN_NE = '0.1';
-
-export const MAX_NE = '999.9';
-
-export const MAX_NE_DECIMAL = 1;
-
-export const INVALID_NE_RANGE_MSG = `Must be between ${MIN_NE} and ${MAX_NE}`;
-
-export const INVALID_NE_DECIMAL_MSG = `Cannot have more than ${MAX_NE_DECIMAL} decimal place`;
